FIX: alterando o tipo do atributo pontuação na model Participação

+++ src/main/java/com/gs/ecoocean/dto/participacao/ParticipacaoResponseDTO.java
@@ -1,5 +1,7 @@
 package com.gs.ecoocean.dto.participacao;
 
+import java.math.BigDecimal;
+
 import com.gs.ecoocean.dto.partida.PartidaResponseDTO;
 import com.gs.ecoocean.dto.voluntario.VoluntarioResponseDTO;
 import com.gs.ecoocean.model.Participacao;
@@ -7,7 +9,7 @@ import com.gs.ecoocean.model.Participacao;
 public record ParticipacaoResponseDTO(
     VoluntarioResponseDTO voluntario,
     PartidaResponseDTO partida,
-    Integer pontuacao
+    BigDecimal pontuacao
 ) {
 
     public ParticipacaoResponseDTO(Participacao participacao){
